feat: Refactor SignUpForm into subcomponents
