+++ src/components/register/sign-up/SignUp.css
@@ -1,0 +1,30 @@
+.sign-up-section {
+    display: flex;
+    flex-direction: column;
+    gap: 40px;
+    width: 370px;
+}
+
+.sign-up-section header {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+    text-align: start;
+}
+
+.sign-up-section header h1 {
+    font-size: 40px;
+    color: #000;
+    font-weight: 600;
+}
+
+.sign-up-section header p {
+    font-size: 16px;
+    color: #6C7275;
+    font-weight: 400;
+}
+.sign-up-section header p a {
+    font-size: 16px;
+    color: #38CB89;
+    text-decoration: none;
+}
