+++ src/components/register/sign-up/signUpForm/signUpForm.css
@@ -1,79 +1,5 @@
-.sign-up-section {
-    display: flex;
-    flex-direction: column;
-    gap: 40px;
-    width: 370px;
-}
-
-.sign-up-section header {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    text-align: start;
-}
-
-.sign-up-section header h1 {
-    font-size: 40px;
-    color: #000;
-    font-weight: 600;
-}
-
-.sign-up-section header p {
-    font-size: 16px;
-    color: #6C7275;
-    font-weight: 400;
-}
-.sign-up-section header p a {
-    font-size: 16px;
-    color: #38CB89;
-    text-decoration: none;
-}
-
 .sign-up-section form {
     display: flex;
     flex-direction: column;
     gap: 32px;
-}
-
-.sign-up-section form input {
-    border: none;
-    border-bottom: 1px solid  #00000066;
-    height: 32px;
-    color:  #00000066;
-    font-size: 16px;
-    padding-bottom: 10px;
-
-    &::placeholder {
-        color:  #00000066;
-        font-size: 16px;
-    }
-    &:focus-visible {
-        outline: none;
-        outline-offset: 0;
-    }
-}
-
-.terms-container {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.terms-container input {
-    width: 24px;
-    height: 24px;
-}
-
-.terms-container label {
-    font-size: 16px;
-    color: #6C7275;
-    font-weight: 400;
-    display: flex;
-    gap: 3px;
-}
-
-
-.terms-container label a {
-    color: #141718;
-    font-weight: 600;
 }
