+++ src/pages/signUpPage/SignUpPage.jsx
@@ -1,12 +1,12 @@
-import SignUpForm from "../../components/register/sign-up/signUpForm/SignUpForm"
 import ImageBanner from "../../components/register/imageBanner/ImageBanner"
 import "./SignUpPage.css"
+import SignUp from "../../components/register/sign-up/SignUp"
 
 const SignUpPage = () => {
   return (
     <main>
       <ImageBanner/>
-      <SignUpForm/>
+      <SignUp/>
     </main>
   )
 }
